+++ vkworkspace/utils/__init__.py
@@ -1,3 +1,10 @@
 from .keyboard import InlineKeyboardBuilder, InlineKeyboardButton
+from .text import html, md, split_text
 
-__all__ = ["InlineKeyboardBuilder", "InlineKeyboardButton"]
+__all__ = [
+    "InlineKeyboardBuilder",
+    "InlineKeyboardButton",
+    "html",
+    "md",
+    "split_text",
+]